Moved schema changes out of bankAccount and into bankInformation
