+++ src/applications/edu-benefits/0994/pages/bankInformation.js
@@ -1,11 +1,13 @@
 import _ from 'lodash';
+import React from 'react';
 import fullSchema from 'vets-json-schema/dist/22-0994-schema.json';
-import bankAccountUI from 'platform/forms/definitions/bankAccount';
+import oldBankAccountUI from 'platform/forms/definitions/bankAccount';
 import ReviewCardField from 'platform/forms-system/src/js/components/ReviewCardField';
+import environment from 'platform/utilities/environment';
+import { isValidRoutingNumber } from 'platform/forms/validations';
 import { hasNewBankInformation, hasPrefillBankInformation } from '../utils';
 import PaymentView from '../components/PaymentView';
 import PaymentReviewView from '../components/PaymentReviewView';
-import environment from 'platform/utilities/environment';
 
 import {
   bankInfoDescriptionWithInfo,
@@ -17,22 +19,83 @@ import {
 const { bankAccount } = fullSchema.properties;
 
 const hasNewBankInfo = formData => {
-  const bankAccountObj = _.get(formData['view:bankAccount'], 'bankAccount', {});
-  return hasNewBankInformation(bankAccountObj);
+  return hasNewBankInformation(formData);
 };
 
 export const hasPrefillBankInfo = formData => {
-  const bankAccountObj = _.get(formData, 'prefillBankAccount', {});
-  return hasPrefillBankInformation(bankAccountObj);
+  return hasPrefillBankInformation(formData);
 };
 
 const startInEdit = data =>
   !_.get(data, 'view:hasBankInformation', false) &&
   !hasNewBankInformation(data.bankAccount);
 
-const newItemName = environment.isProduction()
-  ? 'account'
-  : 'account information';
+const isProduction = environment.isProduction();
+
+const newItemName = isProduction ? 'account' : 'account information';
+
+const directDepositDescription = () => {
+  return isProduction ? (
+    ''
+  ) : (
+    <div className="vads-u-margin-top--2 vads-u-margin-bottom--2">
+      <p>
+        We make payments only through direct deposit, also called electronic
+        funds transfer (EFT). Please provide your direct deposit information
+        below. We’ll send your housing payment to this account.
+      </p>
+      <img
+        src="/img/direct-deposit-check-guide.svg"
+        alt="On a personal check, find your bank’s 9-digit routing number listed along the bottom-left edge, and your account number listed beside that."
+      />
+    </div>
+  );
+};
+
+function validateRoutingNumber(
+  errors,
+  routingNumber,
+  formData,
+  schema,
+  errorMessages,
+) {
+  if (!isValidRoutingNumber(routingNumber)) {
+    errors.addError(errorMessages.pattern);
+  }
+}
+
+const newUiSchema = {
+  'ui:order': ['accountType', 'routingNumber', 'accountNumber'],
+  'ui:description': directDepositDescription(),
+  accountType: {
+    'ui:title': 'Account type',
+    'ui:widget': 'radio',
+    'ui:options': {
+      labels: {
+        checking: 'Checking',
+        savings: 'Savings',
+      },
+    },
+  },
+  accountNumber: {
+    'ui:title': 'Bank account number',
+    'ui:errorMessages': {
+      pattern: 'Please enter a valid account number',
+      required: 'Please enter a bank account number',
+    },
+  },
+  routingNumber: {
+    'ui:title': 'Bank’s 9 digit routing number',
+    'ui:validations': [validateRoutingNumber],
+    'ui:required': true,
+    'ui:errorMessages': {
+      pattern: 'Please enter a valid 9 digit routing number',
+      required: 'Please enter a routing number',
+    },
+  },
+};
+
+const bankAccountUI = isProduction ? oldBankAccountUI : newUiSchema;
 
 export const uiSchema = {
   'ui:title': 'Direct deposit information',
